[이상아] 숙소 관리 시스템 hotel entity 작성 완료 [HMS]
