+++ backend/src/main/java/com/example/demo/entity/Room/Room.java
@@ -1,0 +1,14 @@
+package com.example.demo.entity.Room;
+
+import lombok.Data;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+@Data
+@Entity
+@Table( name = "Room" )
+public class Room {
+}
