+++ backend/src/main/java/com/example/demo/entity/Room/RoomInfo.java
@@ -1,0 +1,4 @@
+package com.example.demo.entity.Room;
+
+public class RoomInfo {
+}
