+++ backend/src/main/java/com/example/demo/entity/Room/RoomType.java
@@ -1,0 +1,10 @@
+package com.example.demo.entity.Room;
+
+import lombok.Data;
+
+import javax.persistence.Entity;
+
+@Data
+@Entity
+public class RoomType {
+}
