+++ backend/src/main/java/com/example/demo/service/member/CustomOAuth2UserService.java
@@ -4,7 +4,6 @@ package com.example.demo.service.member;
 import com.example.demo.entity.member.Role;
 import com.example.demo.entity.member.User;
 import com.example.demo.repository.member.UserRepository;
-import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
@@ -17,10 +16,8 @@ import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 
 import javax.servlet.http.HttpSession;
-import java.math.BigInteger;
 import java.util.Map;
 import java.util.Optional;
-import java.util.OptionalInt;
 import java.util.stream.Collectors;
 
 @Slf4j
