+++ frontend/src/router/index.js
@@ -2,6 +2,8 @@ import Vue from 'vue'
 import VueRouter from 'vue-router'
 import HomeView from '../views/HomeView.vue'
 
+
+
 Vue.use(VueRouter)
 
 const routes = [
